+++ doc/version-history.rst
@@ -9,6 +9,7 @@ v2.8.0
 
 * Communicate with two flow meters.
 * Save and load glycol/heaters setpoints, log errors occuring during saving.
+* Controller::set_setpoints save parameter.
 * Read flow meter identification.
 * Catch error in VFD/EGW Pump read after error.